# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Configuration Management API",
    "version": "v1",
    "description": "Multi-tenant configuration management with versioning and access control"
  },
  "paths": {
...
    "/api/v1/audit/": {
      "get": {
        "operationId": "list_global_audit",
        "summary": "List Global Audit",
        "parameters": [
          {
            "in": "query",
            "name": "auth_id",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auth Id"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "auth_email",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auth Email"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "auth_type",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auth Type"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "object_type",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Type"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "object_id",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Id"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "http_method",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Http Method"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "api_endpoint",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Endpoint"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "permission_ok",
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Permission Ok"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "resolve_type",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resolve Type"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "from_cache",
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Cache"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "event_kind",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Kind"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "parent_event_id",
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Event Id"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "client_ip",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Ip"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "user_agent",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Agent"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "token_number",
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token Number"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "object_version",
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Version"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "operation",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "subresource_type",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subresource Type"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "subresource",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subresource"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "event_id",
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Id"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "error",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "namespace",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAuditEventSchema"
                }
              }
            }
          }
        },
        "description": "Cross-namespace audit log (global admin only).",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/audit/{event_id}": {
      "get": {
        "operationId": "get_global_audit_event",
        "summary": "Get Global Audit Event",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "schema": {
              "title": "Event Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Get a single audit event (global admin only).",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
...
    "/api/v1/ns/{namespace}/~audit/timeline/": {
      "get": {
        "operationId": "namespace_audit_timeline",
        "summary": "Namespace Audit Timeline",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "object_id",
            "schema": {
              "title": "Object Id",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "object_type",
            "schema": {
              "title": "Object Type",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAuditTimelineEntrySchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Item-scoped audit timeline (requires ``<object_type>:audit`` on ``object_id``).",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~audit/resolve-series/": {
      "get": {
        "operationId": "namespace_audit_resolve_series",
        "summary": "Namespace Audit Resolve Series",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "object_id",
            "schema": {
              "title": "Object Id",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "object_type",
            "schema": {
              "title": "Object Type",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "bucket_seconds",
            "schema": {
              "minimum": 1800,
              "title": "Bucket Seconds",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveSeriesSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Time-bucketed direct vs nested resolve counts for a tree object.",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~audit/{event_id}": {
      "get": {
        "operationId": "get_namespace_audit_event",
        "summary": "Get Namespace Audit Event",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "event_id",
            "schema": {
              "title": "Event Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Get a single audit event within a namespace.",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorSchema": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Error"
          },
          "audit_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Audit log event ID when the failure was recorded server-side",
            "title": "Audit Event Id"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorSchema",
        "type": "object"
      },
...
      "AuditEventSchema": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "client_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Ip"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "auth_id": {
            "title": "Auth Id",
            "type": "string"
          },
          "auth_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Email"
          },
          "auth_type": {
            "title": "Auth Type",
            "type": "string"
          },
          "token_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Number"
          },
          "namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Denormalized namespace name",
            "title": "Namespace"
          },
          "http_method": {
            "title": "Http Method",
            "type": "string"
          },
          "api_endpoint": {
            "title": "Api Endpoint",
            "type": "string"
          },
          "object_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Type"
          },
          "object_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Id"
          },
          "object_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Version"
          },
          "operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation"
          },
          "subresource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subresource Type"
          },
          "subresource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subresource"
          },
          "permission_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "resolve_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolve Type"
          },
          "from_cache": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Cache"
          },
          "parent_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Event Id"
          },
          "event_kind": {
            "title": "Event Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "occurred_at",
          "auth_id",
          "auth_type",
          "http_method",
          "api_endpoint",
          "event_kind"
        ],
        "title": "AuditEventSchema",
        "type": "object"
      },
      "PagedAuditEventSchema": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedAuditEventSchema",
        "type": "object"
      },
...
      "PagedAuditTimelineEntrySchema": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditTimelineEntrySchema"
            },
            "title": "Items",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedAuditTimelineEntrySchema",
        "type": "object"
      },
      "ResolveSeriesBucketSchema": {
        "properties": {
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "direct": {
            "default": 0,
            "title": "Direct",
            "type": "integer"
          },
          "nested": {
            "default": 0,
            "title": "Nested",
            "type": "integer"
          },
          "errors": {
            "default": 0,
            "title": "Errors",
            "type": "integer"
          }
        },
        "required": [
          "start"
        ],
        "title": "ResolveSeriesBucketSchema",
        "type": "object"
      },
      "ResolveSeriesSchema": {
        "properties": {
          "bucket_seconds": {
            "title": "Bucket Seconds",
            "type": "integer"
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/ResolveSeriesBucketSchema"
            },
            "title": "Buckets",
            "type": "array"
          }
        },
        "required": [
          "bucket_seconds",
          "buckets"
        ],
        "title": "ResolveSeriesSchema",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ResolverAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "openid": "",
              "profile": "",
              "email": "",
              "groups": ""
            },
            "authorizationUrl": "http://localhost:8080/dex/auth",
            "tokenUrl": "http://localhost:8080/dex/token"
          }
        }
      }
    }
  },
  "servers": []
}